backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
pillow==10.2.0
pypdf2==3.0.1
pytesseract==0.3.10
google-generativeai==0.3.2
openai==1.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
